abstracted out the scim config from org attributes into a full config interface.
