+++ gradle.properties
@@ -3,7 +3,7 @@ junitVersion=5.11.3
 keycloakVersion=26.1.4
 org.gradle.configuration-cache=true
 testContainersKeycloakVersion=3.6.0
-testContainersVersion=1.20.4
+testContainersVersion=1.21.4
 awaitilityVersion=4.3.0
 seleniumRemoteDriverVersion=4.26.0
 seleniumVersion=4.26.0
