+++ src/main/java/fi/metatavu/keycloak/scim/server/organization/OrganizationScimConfig.java
@@ -8,7 +8,7 @@ import java.util.Map;
 /**
  * SCIM configuration for organizations
  */
-public abstract class OrganizationScimConfig implements ScimConfig {
+public interface OrganizationScimConfig extends ScimConfig {
 
     public static final String SCIM_EXTERNAL_SHARED_SECRET = "SCIM_EXTERNAL_SHARED_SECRET";
     public static final String SCIM_EXTERNAL_JWKS_URI = "SCIM_EXTERNAL_JWKS_URI";
@@ -18,8 +18,7 @@ public abstract class OrganizationScimConfig implements ScimConfig {
     public static final String SCIM_AUTHENTICATION_MODE = "SCIM_AUTHENTICATION_MODE";
     public static final String SCIM_EMAIL_AS_USERNAME = "SCIM_EMAIL_AS_USERNAME";
 
-    @Override
-    public void validateConfig() throws ConfigurationError {
+    default void validateConfig() throws ConfigurationError {
         AuthenticationMode mode = getAuthenticationMode();
         if (mode == null) {
             throw new ConfigurationError(SCIM_AUTHENTICATION_MODE + " is not set");
@@ -51,46 +50,9 @@ public abstract class OrganizationScimConfig implements ScimConfig {
         }
     }
 
-    @Override
-    public AuthenticationMode getAuthenticationMode() {
-        String value = getAttribute(SCIM_AUTHENTICATION_MODE);
-        if (value == null || value.isEmpty()) {
-            return null;
-        }
-
-        return AuthenticationMode.valueOf(value);
-    }
-
-    @Override
-    public String getExternalIssuer() {
-        return getAttribute(SCIM_EXTERNAL_ISSUER);
-    }
-
-    @Override
-    public String getExternalJwksUri() {
-        return getAttribute(SCIM_EXTERNAL_JWKS_URI);
-    }
-
-    @Override
-    public String getExternalAudience() {
-        return getAttribute(SCIM_EXTERNAL_AUDIENCE);
-    }
-
-    @Override
-    public String getSharedSecret() {
-        return getAttribute(SCIM_EXTERNAL_SHARED_SECRET);
-    }
-
-    @Override
-    public boolean getLinkIdp() {
-        return "true".equalsIgnoreCase(getAttribute(SCIM_LINK_IDP));
-    }
-
-    @Override
-    public boolean getEmailAsUsername() {
-        return "true".equalsIgnoreCase(getAttribute(SCIM_EMAIL_AS_USERNAME));
-    }
-
-    public abstract String getAttribute(String attributeName);
+    /**
+     * Is the organization enabled for SCIM
+     */
+    public boolean isEnabled();
 
 }
